Organize client components by directory
