+++ packages/client/components/FollowerList/index.ts
@@ -1,0 +1,2 @@
+export { FollowerList } from './FollowerList';
+export { default } from './FollowerList';
