+++ packages/client/components/FollowersPanel/index.ts
@@ -1,0 +1,2 @@
+export { FollowersPanel } from './FollowersPanel';
+export { default } from './FollowersPanel';
