+++ packages/client/components/GroupMembers/GroupMembers.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { GroupMembers } from './GroupMembers';
-import { sampleGroup } from '../lib/data';
+import { sampleGroup } from '../../lib/data';
 
 const meta: Meta<typeof GroupMembers> = {
   title: 'Groups/GroupMembers',
