+++ packages/client/components/GroupMembers/index.ts
@@ -1,0 +1,2 @@
+export { GroupMembers } from './GroupMembers';
+export { default } from './GroupMembers';
