+++ packages/client/components/LoginFlow/index.ts
@@ -1,0 +1,2 @@
+export { LoginFlow } from './LoginFlow';
+export { default } from './LoginFlow';
