+++ packages/client/components/ProfileSummary/ProfileSummary.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { ProfileSummary } from './ProfileSummary';
-import { sampleProfile } from '../lib/data';
+import { sampleProfile } from '../../lib/data';
 
 const meta: Meta<typeof ProfileSummary> = {
   title: 'Profile/ProfileSummary',
