+++ packages/client/components/ProfileSummary/index.ts
@@ -1,0 +1,2 @@
+export { ProfileSummary } from './ProfileSummary';
+export { default } from './ProfileSummary';
