+++ packages/client/components/StatusBar/index.ts
@@ -1,0 +1,2 @@
+export { StatusBar } from './StatusBar';
+export { default } from './StatusBar';
