+++ packages/client/vitest.config.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { defineConfig, defineProject } from 'vitest/config';
 import react from '@vitejs/plugin-react';
 import { storybookTest } from '@storybook/addon-vitest/vitest-plugin';
